Rescoped Resource records
Split the storage between the 'saved' records and the modification
history for the current session. Also changed it so that only the
current persistent resource should be modifiable.

Moved these new types out of the resource library and intot he
simulation library, with an eye towards moving them to a higher
abstraction layer later.