Add CMake install/configuration capabilities
This commit makes it much easier to make the software to become
available beyond just it's source code form.

All related headers and libraries can now be installed where desired and
into a few basic consistent directories. This should allow for easy
packaging in the future.

As well, this includes CMake configuration files that allow for easy use
of the `find_package` call to find and configure use.