Removed WSI libraries, merged into bringup
The WSI library and it's interface was too awkward to use in it's
current state, and required alot of overhead bouncing back and forth
between the application and the library internals, especially around the
associated Vulkan surfaces.

For now, the WSI libraries have been removed and the functionality
entirely moved into the bringup application.