Fix misalignment issues with crypto keys
The allocation returned by libsodium is not always aligned to perfectly
match the type. As such, increase the complexity by also having a
uint8_t at the initial pointer location to note the offset of the data
and then read the key size, then have the data after a buffer after
that.