Update to swapchain extent sizing
The sizing up until now has been usually set to the current surface
extent, unless it didn't have one, in which case it was set to the input
width/height values. This is confusing and didn't allow for custom
rendering resolutions.

Now, the framebuffer now takes from the input width/height values,
unless they are both zero in which case it grabs the surface extent
and if that fails returns an error.

Also now actually checks against extent min/max.