Update thread pool end notification process
It seems that in rare cases that when the pool is being destroyed, after
the notifications occur the yield may never return, and thus the program
gets into a frozen state.

Rather than yield, sleep the thread for only a milliescond then repeat
the notification process until the threads complete and join.